+++ FormatFactory.Tests/FormatReaderTests.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using Company.Entities;
 using Company.Entities.Malformed;
 using Irvin.FormatFactory;
-using Company.Nacha.NachaElements;
 using NUnit.Framework;
 using InvalidDataException = Irvin.FormatFactory.InvalidDataException;
 
@@ -15,22 +14,6 @@ namespace TestProject
     public class FormatReaderTests
     {
         private const string NON_DATA_FILE = @"c:\windows\system32\drivers\etc\hosts";
-        private string _temporaryFilename;
-
-        [TestFixtureSetUp]
-        public void RunFirstOnce()
-        {
-            _temporaryFilename = Path.GetTempFileName();
-        }
-
-        [SetUp]
-        public void RunBeforeEachTest()
-        {
-            if (File.Exists(_temporaryFilename))
-            {
-                File.Delete(_temporaryFilename);
-            }
-        }
 
         [TestCase(typeof(Person))]
         [TestCase(typeof(Person2))]
@@ -212,18 +195,6 @@ namespace TestProject
         }
 
         [Test]
-        public void Read_ParsesEnumValue_AccordingToSpecifier()
-        {
-            Thing element = new Thing();
-            element.Goober = "ABCD";
-            element.CompareKind = StringComparison.Ordinal;
-            element.C2 = StringComparison.CurrentCultureIgnoreCase;
-            element.C3 = StringComparison.InvariantCulture;
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
         public void Read_CanParseBareCsvFormat()
         {
             string input = "toad,\"gargalong\"" + Environment.NewLine + "\"zupa, toscana\",fletermouse";
@@ -289,59 +260,6 @@ namespace TestProject
         }
 
         [Test]
-        public void Read_ReadsBackValuesForMinMaxLengthFields_RegardlessOfFixedWidthSettings()
-        {
-            GoodD element = new GoodD();
-            element.Gorzo = "fa";
-            element.JackieChan = "la";
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
-        public void Read_ReadsValuesForMinMaxLengthFields_RegardlessOfFixedWidthSettings_ForValueThatsBetweenMinimumAndMaximum()
-        {
-            GoodD element = new GoodD();
-            element.Gorzo = "fafetnu";
-            element.JackieChan = "la";
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
-        public void Read_ReadsValuesForMinMaxLengthFields_ForFixedMinWidthOnly()
-		{
-			GoodD element = new GoodD();
-			element.Gorzo = "fafetnu";
-			element.JackieChan = "la";
-			element.WWF = true;
-
-            TestWriteReadCycle(element);
-		}
-
-        [Test]
-        public void Read_ReadsCorrectly_ForMixOfAccess()
-        {
-            GoodE element = new GoodE();
-            element.Gorzo = "fafetnu";
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
-        public void Read_ReadsDesiredResult_ForFixedWidthAttributeFields()
-        {
-            InterchangeControlLoop element = new InterchangeControlLoop();
-            element.InterchangeSenderID = 615156491;
-            element.InterchangeDate = new DateTime(2008, 1, 25);
-            element.InterchangeTime = DateTime.Now.Date.AddHours(7).AddMinutes(38);
-            element.InterchangeControlNumber = 4;
-            element.UsageIndicator = 'T';
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
         [ExpectedException(typeof(InvalidUsageException), ExpectedMessage = "A member cannot be a field and a sub-record (violator: 'Name')")]
         public void Read_ThrowsException_IfFieldAndSubRecordsCombined()
         {
@@ -380,31 +298,6 @@ namespace TestProject
         }
 
         [Test]
-        public void Read_ReadsRecordsInCorrectOrder_WithMixOfFieldsAndProperties()
-        {
-            Stats element = new Stats();
-            element.Name = "greg";
-            element.FigureA = 34;
-            element.FigureB = 45;
-            element.FigureD = 99;
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
-        public void Read_ReadsRecordWithWeirdRules()
-        {
-            ProfessionalServiceLoop element = new ProfessionalServiceLoop();
-            element.ProcedureCodeParts.ProcedureCode = "S0500";
-            element.LineItemChargeAmount = 89.25m;
-            element.ServiceUnitCount = 1;
-            element.DiagnosisCodePointer1 = 1;
-            element.DiagnosisCodePointer2 = 1;
-
-            TestWriteReadCycle(element);
-        }
-
-        [Test]
         public void Read_ParsesDataCorrectly_IfNamesInUnusualOrder_WhenNoSpecifiedOrder()
         {
             string content = "Total,OrderNumber,Ordered,ShipName\r\n36,ABCD1,12/31/2007,Tom Hardy 1\r\n";
@@ -427,80 +320,6 @@ namespace TestProject
         }
 
         [Test]
-        public void Read_ReadsStandardCsvRecord()
-        {
-            Order subject = OrderTestHelper.GetOrder();
-            subject.IsActive = false;
-
-            TestWriteReadCycle(subject);
-        }
-
-        [Test]
-        public void Read_ReadsRecordWithOddDecorationOrder()
-        {
-            OrderQ subject = OrderTestHelper.GetOrderQ();
-
-            TestWriteReadCycle(subject);
-        }
-
-        [TestCase(true)]
-        [TestCase(false)]
-        public void Read_ReadsHeaderlessCsvRecord_RegardlessOfHeaderSetting(bool headers)
-        {
-            Order subject = OrderTestHelper.GetOrder();
-            subject.IsActive = false;
-            List<Order> source = new List<Order> {subject};
-            FormatWriter.Instance.Write(_temporaryFilename, source);
-            FormatOptions options = new FormatOptions
-            {
-                FieldDelimiter = ",",
-                IncludeHeaders = headers
-            };
-
-            List<Order> actual = FormatReader.Default.ReadFromFile<Order>(_temporaryFilename, options);
-
-            ObjectAssert.ListsEqual(source, actual);
-        }
-
-        [TestCase(true)]
-        [TestCase(false)]
-        public void Read_ReadsCsvRecord_RegardlessOfHeaderSetting_ForComprehensiveAttribute(bool headers)
-        {
-            OrderB subject = OrderTestHelper.GetOrderB();
-            subject.IsActive = false;
-            List<OrderB> source = new List<OrderB> { subject };
-            FormatWriter.Instance.Write(_temporaryFilename, source);
-            FormatOptions options = new FormatOptions
-            {
-                FieldDelimiter = ",",
-                IncludeHeaders = headers
-            };
-
-            List<OrderB> actual = FormatReader.Instance.ReadFromFile<OrderB>(_temporaryFilename, options);
-
-            ObjectAssert.ListsEqual(source, actual);
-        }
-
-        [TestCase(true)]
-        [TestCase(false)]
-        public void Read_ReadsCsvRecord_RegardlessOfHeaderSetting_ForComprehensiveAttribute_UnusualOrder(bool headers)
-        {
-            OrderC subject = OrderTestHelper.GetOrderC();
-            subject.IsActive = false;
-            List<OrderC> source = new List<OrderC> { subject };
-            FormatWriter.Instance.Write(_temporaryFilename, source);
-            FormatOptions options = new FormatOptions
-            {
-                FieldDelimiter = ",",
-                IncludeHeaders = headers
-            };
-
-            List<OrderC> actual = FormatReader.Default.ReadFromFile<OrderC>(_temporaryFilename, options);
-
-            ObjectAssert.ListsEqual(source, actual);
-        }
-
-        [Test]
         public void Read_ReadsCsvCorrectly_ForOddDecorationOrder()
         {
             string content = "OrderNumber,ShippingName,Ordered,OrderTotal\r\n5,bob,11/11/2017,19.99";
@@ -511,81 +330,6 @@ namespace TestProject
             Assert.AreEqual("bob", actual.ShippingName);
             Assert.AreEqual(new DateTime(2017, 11, 11).Date, actual.Ordered.Date);
             Assert.AreEqual(19.99m, actual.OrderTotal);
-        }
-
-        [Test]
-        public void Read_ReadsStandardCsvRecord_WithHeaders()
-        {
-            Order2 subject = OrderTestHelper.GetOrder2();
-            subject.IsActive = false;
-
-            TestWriteReadCycle(subject);
-        }
-
-        [Test]
-        public void Read_ReadsStandardCsvRecord_WithHeaders_OrderSpecified()
-        {
-            Order3 subject = OrderTestHelper.GetOrder3();
-            subject.IsActive = false;
-
-            TestWriteReadCycle(subject);
-        }
-
-        [TestCase(true)]
-        [TestCase(false)]
-        public void Read_ReadsCsvRecord_RegardlessOfHeaderSetting(bool headers)
-        {
-            Order2 subject = OrderTestHelper.GetOrder2();
-            subject.IsActive = false;
-            List<Order2> source = new List<Order2> { subject };
-            FormatWriter.Instance.Write(_temporaryFilename, source);
-            FormatOptions options = new FormatOptions
-            {
-                FieldDelimiter = ",",
-                IncludeHeaders = headers
-            };
-
-            List<Order2> actual = FormatReader.Instance.ReadFromFile<Order2>(_temporaryFilename, options);
-
-            ObjectAssert.ListsEqual(source, actual);
-        }
-
-        [Test]
-        public void Read_ReadsRecords()
-        {
-            List<Order> list = new List<Order>
-				{
-					OrderTestHelper.GetOrder(),
-					OrderTestHelper.GetOrder(2)
-				};
-            list.ForEach(x => x.IsActive = false);
-
-            TestWriteReadCycle(list);
-        }
-
-        [TestCase(1)]
-        [TestCase(12)]
-        [TestCase(123)]
-        [TestCase(1234)]
-        [TestCase(123456)]
-        public void Reads_ReadsRecordWithMinimumWidth(int value)
-        {
-            Stats item = new Stats();
-            item.Id = value;
-
-            TestWriteReadCycle(item);
-        }
-
-        [Test]
-        public void Read_ReadsPipeDelimitedRecord()
-        {
-            LegacyOrder subject = new LegacyOrder();
-            subject.OrderNumber = "ABCD";
-            subject.OrderTotal = 35m;
-            subject.Ordered = new DateTime(2007, 12, 31);
-            subject.ShippingName = "Tom Hardy";
-
-            TestWriteReadCycle(subject);
         }
 
         [Test]
@@ -675,31 +419,6 @@ namespace TestProject
             expected[0].Children.Add(new SimpleFooter { FirstName = "Tom Jr.", LastName = "Hardy", Age = 8 });
 
             FormatReader.Default.Read<SimpleHeader>(input);
-        }
-
-        [Test]
-        public void Read_DoesNotRunOffTheEndOfString_IfContentAllPadding()
-        {
-            FileHeaderRecord record = NachaFileTestHelper.GetFileHeaderRecord();
-            record.ImmediateDestination = "          ";
-
-            TestWriteReadCycle(record);
-        }
-        
-        private void TestWriteReadCycle<T>(T element)
-            where T : class
-        {
-            TestWriteReadCycle(new List<T> {element});
-        }
-
-        private void TestWriteReadCycle<T>(List<T> source)
-            where T : class
-        {
-            FormatWriter.Instance.Write(_temporaryFilename, source);
-
-            List<T> actual = FormatReader.Default.ReadFromFile<T>(_temporaryFilename);
-
-            ObjectAssert.ListsEqual(source, actual);
         }
     }
 }
